Finish expedition to zram
Tried to setup BLOCKDEVICE -> LUKS -> GPT -> BLOCKDEVICE**
where ** is the backing store for zram, but encountered issues with
sector sizes => the host didn't want to boot.
Didn't dug in to find out why stage-1-boot isn't happy with that setup
and replicated the disko example of LVM on top of LUKS. The end situation
is not BLOCKDEVICE -> LUKS -> LVM -> BLOCKDEVICE**

Also tweaked some kernel params and setup a zram management service.